payouts functionality decoupled
