+++ src/v2/constants/safe/gnosis_safe.tsx
@@ -3,8 +3,10 @@ import Safe, { ContractNetworksConfig } from '@gnosis.pm/safe-core-sdk'
 import EthersAdapter from '@gnosis.pm/safe-ethers-lib'
 import SafeServiceClient from '@gnosis.pm/safe-service-client'
 import axios from 'axios'
-import { ethers } from 'ethers'
+import { ethers, logger } from 'ethers'
 import { MetaTransaction, Safe as _GnosisSafe, TransactionType } from 'src/v2/types/safe'
+import { getCeloTokenUSDRate, loadAssetId } from 'src/v2/utils/tokenToUSDconverter'
+import { erc20ABI } from 'wagmi'
 
 export class GnosisSafe implements _GnosisSafe {
 	id: string
@@ -36,9 +38,61 @@ export class GnosisSafe implements _GnosisSafe {
 		throw new Error('Method not implemented.')
 	}
 
-	async createMultiTransaction(transactions: MetaTransaction[], safeAddress: string) {
+	encodeTransactionData(recipientAddress: string, fundAmount: string, rewardAssetDecimals: number) {
+		const ERC20Interface = new ethers.utils.Interface(erc20ABI)
+		const txData = ERC20Interface.encodeFunctionData('transfer', [
+			recipientAddress,
+			ethers.utils.parseUnits(fundAmount, rewardAssetDecimals)
+		])
 
-		console.log('creating gnosis transaction for', transactions)
+		return txData
+	}
+
+	async createEVMMetaTransactions(workspaceSafeChainId: string , gnosisBatchData: any): Promise<MetaTransaction[]> {
+
+		const celoTokensUSDRateMapping = await (await getCeloTokenUSDRate()).data;
+		const readyTxs = gnosisBatchData.map((data: any) => {
+			let tokenUSDRate: number
+			if(workspaceSafeChainId === '42220') {
+				const tokenSelected = data.selectedToken.name.toLowerCase()
+				if(tokenSelected === 'cusd') {
+					tokenUSDRate = celoTokensUSDRateMapping['celo-dollar'].usd
+				} else if(tokenSelected === 'ceuro') {
+					tokenUSDRate = celoTokensUSDRateMapping['celo-euro'].usd
+				} else if(tokenSelected === 'tether') {
+					tokenUSDRate = celoTokensUSDRateMapping['tether'].usd
+				} else if(tokenSelected === 'spcusd') {
+					tokenUSDRate = 1
+				} else if(tokenSelected === 'spCELO') {
+					tokenUSDRate = 1
+				}
+			} else {
+				tokenUSDRate = data.selectedToken.info.fiatConversion
+			}
+
+			const rewardAssetDecimals = data.selectedToken.info.decimals
+			const rewardAssetAddress = data.selectedToken.info.tokenAddress
+			const usdToToken = (data.amount / tokenUSDRate!).toFixed(rewardAssetDecimals)
+
+			// console.log('reward asset address', rewardAssetAddress)
+			logger.info('usd amount, usd rate, usd to token amount', data.amount, tokenUSDRate!, usdToToken)
+			const txData = this.encodeTransactionData(data.to, (usdToToken.toString()), rewardAssetDecimals)
+			const tx = {
+				to: ethers.utils.getAddress(rewardAssetAddress),
+				data: txData,
+				value: '0'
+			}
+			return tx
+		})
+
+		return readyTxs
+	}
+
+	async createMultiTransaction(workspaceSafeChainId: any, initiateTransactionData: any, safeAddress: string) {
+
+		const readyToExecuteTxs = await this.createEVMMetaTransactions(workspaceSafeChainId, initiateTransactionData)
+
+		console.log('creating gnosis transaction for', readyToExecuteTxs)
 		//@ts-ignore
 		const provider = new ethers.providers.Web3Provider(window.ethereum)
 		await provider.send('eth_requestAccounts', [])
@@ -71,7 +125,7 @@ export class GnosisSafe implements _GnosisSafe {
 		}
 
 		try {
-			const safeTransaction = await safeSdk.createTransaction({safeTransactionData: transactions})
+			const safeTransaction = await safeSdk.createTransaction({safeTransactionData: readyToExecuteTxs})
 
 			const safeTxHash = await safeSdk.getTransactionHash(safeTransaction)
 			const senderSignature = await safeSdk.signTransactionHash(safeTxHash)
