+++ src/v2/constants/safe/realms_solana.tsx
@@ -450,7 +450,7 @@ const getTokenAndbalance = async(realmAddress: string): Promise<any> =>{
 	])
 	const governance = governances.filter((gov)=>gov.pubkey.toString()===realmData.account.authority?.toString())[0]
 	const nativeTreasuryAddress = await getNativeTreasuryAddress(programId, governance.pubkey)
-	assert(realmData.account.name)
+	// assert(realmData.account.name)
 	const solAmount = (await connection.getAccountInfo(nativeTreasuryAddress))!.lamports / 1000000000
 	const usdAmount = await solanaToUsd(solAmount)
